Produces density plots of parameter estimates from the fitted models

Produces a ggplot2 object so can be further customised as desired. If the results of a node-splitting analysis is supplied, the output is a list of plots, one for each analysis and one for the consistency fit.

plotPosteriorDensities(
  posteriorDraws,
  separateChains = TRUE,
  lineAlpha = 0.7,
  lineSize = 0.2,
  densityResolution = 256,
  wrapByModelName = FALSE
)

Arguments

posteriorDraws
tidy data frame, all posterior draws with (at least) the following columns: modelName, chainId, mcmcIterationId, parameterName, parameterValue.

separateChains
boolean, should chains be plotted as separate distributions (TRUE, default) or as one (FALSE)? Not plotting chains separately somewhat defeats the purpose of the visualisation but is anyway left to the user.

lineAlpha
the opacity of drawn lines. If alpha = 1, lines will be entirely opaque masking overlain lines, so the default is 0.7.

lineSize
scalar, thickness of the plot lines.

densityResolution
scalar, the number of grid points over which to estimate the densities. Higher values yield smoother curves but produces a larger object.

wrapByModelName
boolean, if the modelName column contains more than one distinct value, this should be set to TRUE to show separate plots for each model. Otherwise, the plot will combine samples for the same parameter from different models, which will give misleading results.

Value

A ggplot object, or a list of ggplot objects.