Runs the analysis and returns results ready to be stored in database or used directly

Runs the analysis and returns results ready to be stored in database or used directly

runAnalyses(
  aggregatedResults,
  includeNodesplittingAnalysis = FALSE,
  nCores = 4,
  nWarmup = 1000,
  nIter = 3000,
  nChains = 4,
  fitThin = 1,
  referenceTreatmentId = NULL,
  preferredDirection = 1,
  modelType = "random",
  drawsThin = 10,
  nmaIds = NULL
)

Arguments

aggregatedResults
tidy data frame, result of calling fetchTcoEstimates.

includeNodesplittingAnalysis
boolean, should also node-splitting analyses be run? NB: This can be quite time-consuming, so is by default set to FALSE.

nCores
scalar, how many cores to use when running the analyses. If > 1, foreach is used to parallel processing.

nWarmup
scalar, the number of iterations the sampler should do before it starts sampling from the posterior. This should be sufficient to the chains to have converged.

nIter
scalar, the number of posterior samples to draw per chain.

nChains
scalar, the number of JAGS chains to run

fitThin
scalar, thinning factor. 1 (default) keeps all samples whereas e.g. 10 would keep only every tenth sample.

referenceTreatmentId
string, the name of the exposure to use as the reference when computing pairwise estimates. If none given, the first in alphabetical order will be used.

preferredDirection
scalar, preferential direction of the outcome. 1 means higher values are preferred, -1 means lower values are preferred.

modelType
string defining whether to run a random-effects ("random", default) or a fixed-effects ("fixed") mode.

drawsThin,
scalar, thinning factor used when fetching posterior draws to keep in the database. Used to produce further derived quantities, density plots of posterior distributions and trace plots.

nmaIds
scalar vector, if only specific network meta analyses should be run, their id's are supplied here. If none supplied, all will be run.